Refactor base.ModuleCollection to recipe.Recipe
