+++ PYME/recipes/__init__.py
@@ -2,3 +2,5 @@
 Recipes - these provide a way of specifying a customizable analysis pipeline as 
 a series of blocks which can be chained together.
 """
+
+from .recipe import Recipe
